Take Test: Question card height stabilised for most questions
This change minimises the likelihood that the question container’s
height changes between questions of different text volumes. I.e. if a
user sees a short question with two options and goes to a mid-length
question with 5 options, the div should remain the same height for
better UX.

This is a double-edged sword, if we account for too much variation, the
card design becomes less elegant - equally, if we focus too much on
design, the user will come across more questions where the div height
is adjusted to fit.