Critical Code Fix: Searching now done via cloud code in between Algolia
This had to be done to attach the test author’s: embedding author’s on
tests when indexing a test is fine, but that object becomes outdated
quickly. Therefore, we need to fetch updated data using cloud code.

This was causing a number of ‘incorrect stats’ issues, as the website
was reverting the ‘currentUser’ to old, outdated data, when received
from Algol.

Long-story short, this will slightly increase searching time, but is
critical.